Changing package structure for stages
