+++ src/main/java/org/albacete/simd/framework/BHCStage.java
@@ -1,8 +1,8 @@
-package org.albacete.simd.algorithms.framework.stages;
+package org.albacete.simd.framework;
 
 import edu.cmu.tetrad.graph.Edge;
 import edu.cmu.tetrad.graph.Graph;
-import org.albacete.simd.threads.BESThread;
+import org.albacete.simd.threads.BackwardsHillClimbingThread;
 import org.albacete.simd.threads.GESThread;
 import org.albacete.simd.utils.Problem;
 import org.albacete.simd.utils.Utils;
@@ -10,13 +10,10 @@ import org.albacete.simd.utils.Utils;
 import java.util.ArrayList;
 import java.util.List;
 
-public class BESStage extends ThreadStage {
+public class BHCStage extends ThreadStage {
 
-    public BESStage(Problem problem, int nThreads, int itInterleaving, List<List<Edge>> subsets) {
-        super(problem, nThreads, itInterleaving, subsets);
-    }
 
-    public BESStage(Problem problem, Graph currentGraph, int nThreads, int itInterleaving, List<List<Edge>> subsets) {
+    public BHCStage(Problem problem, Graph currentGraph, int nThreads, int itInterleaving, List<List<Edge>> subsets) {
         super(problem, currentGraph, nThreads, itInterleaving, subsets);
     }
 
@@ -30,9 +27,9 @@ public class BESStage extends ThreadStage {
         //problem.buildIndexing(currentGraph);
 
         // Rearranging the subsets, so that the BES stage only deletes edges of the current graph.
-        List<List<Edge>> subsets_BES = Utils.split(this.currentGraph.getEdges(), this.nThreads);
+        List<List<Edge>> subsets_BHC = Utils.split(this.currentGraph.getEdges(), this.nThreads);
         for (int i = 0; i < this.nThreads; i++) {
-            this.gesThreads[i] = new BESThread(this.problem, this.currentGraph, subsets_BES.get(i));
+            this.gesThreads[i] = new BackwardsHillClimbingThread(this.problem, this.currentGraph, subsets_BHC.get(i));
         }
 
         // Initializing thread config
